Updates two things in public_controller :
-- in list.txt and list.xml, when creating stack of deprecation / supercession project ids,
detects infinite loop and breaks out of it

-- no longer displays wormbase builds in filenames (eg, WS210) to end user when downloading under
controllers download, get_file, & download_tarball .
(builds are still present in original filenames on server but won't be when user downloads copies).
This DOES retain the build name in the liftover folders, since that one is actually useful & expected to be accurate.

The motivation behind removing build names is that the file names are inconsistently lifted when the contents are lifted,
which is very confusing to users.