Improves delete speed via clearing queue-position
Applies miscellaneous changes to VASTLY improve the speed of deleting projects.

command_controller :
Sets the queue_position attribute to nil whenever a command is run.

pipeline_controller : handles queuing some expands when the upload they are following has a nil position

delete_controller : sets a submission's command's queue positions to nil immediately before the sub is deleted,
in case they were run historically and still have a position.

models/command : allows commands to be compared by queue_order, falling back on id if their position is
nil (instead of assigning them a position).

MOTIVATION :

When a Project object is deleted, so are its dependant commands.
A Command object has acts_as_queue (a renamed version of acts_as_list).
A thing that this implies is that when a Command object is deleted, all the
Command objects later than it in the queue are to be moved up one place forward.
(The exception is if the deleted object is not IN a queue, ie it's position is nil.)
This latter operation, when applied to a database with many many Command objects,
is verrrrry slow (since it's applying to ALL Command objects with positions, not just
those with the "queued" status).

And if more of them happens at once, the chance for deadlock seems rather high, in my
observation.

SO, since queue_position is COMPLETELY irrelevant for commands which are no longer
in the queue, it's best to nuke it from orbit.