Fixes geoid attachment when extracted has subfolders
When the SDRF is in a subfolder, the AttachGeoids tool could not
find the GEOids it had just generated to commit them into the database,
because the code searching for them within extracted was different from
the code to locate the SDRF when generating GEOids.

This fixes that; generated GEOids will now be correctly found.