geoID helper: handles case of wrong # of datums
A submission showed up where a placeholder value for the geoID datum
had been used twice, incorrectly, which meant the # of applied protocols
was fine but there were not enough datums. Added a case to check for this
specifically and use the applied protocols as a guide for datum /
input geoid correspondence.